==> code/01_clean_protein_table.ipynb <==
import pandas as pd

## Reading in 3D proximity to glycosites amino acids

df = pd.read_csv('../data/merge_pdb_swissmod_10and6angstroms.ident25.csv',
                 index_col=0)
print(df.shape)
print('Filtering for having non-NA for AAs in structural window')
df = df.dropna(subset=['struct_window_angstrom']).reset_index(drop=True)
print(df.shape)
df.head()

## Creating a flat table of 1 gene and 1AA per row

proteins = []
aas = []
for i in df.index:
    aa_list = df.loc[i]['struct_window_angstrom_resnum'].split('|')
    protein = df.loc[i]['gene_id']
    for aa in aa_list:
        proteins.append(protein)
        aas.append(aa)

just_p_and_aa = pd.DataFrame([proteins,aas]).transpose()
just_p_and_aa.drop_duplicates().to_csv('../data/protein_and_aa.txt',
                                        header=None,
                                        index=False,
                                        sep='\t')

## Incorporating continuous distance into protein table
