==> code/03_path_vs_benign.ipynb <==
import matplotlib.pyplot as plt
import pandas as pd
import scipy
import seaborn as sns

# Asking the question, are clinvar variants that are in amino acids that are near glycosites more likely to be pathogenic than ones that are not?

df = pd.read_csv('../data/clinvar_glycoproteins.txt',
                 sep='\t',
                 index_col=0)

df['aa_change'] = df['aaref'] + '=>' + df['aaalt']

counts = df.groupby(by=['clinvar_pathogenic','glycosite_proximal'],as_index=False).count()[['clinvar_pathogenic','glycosite_proximal','#chr']]
counts

counts_pivot = counts.pivot(index='clinvar_pathogenic',columns='glycosite_proximal',values='#chr')
counts_pivot.loc['likely benign + benign'] = counts_pivot.loc['likely benign'] + counts_pivot.loc['benign'] 
counts_pivot.loc['likely pathogenic + pathogenic'] = counts_pivot.loc['likely pathogenic'] + counts_pivot.loc['pathogenic'] 

counts_pivot

## Testing all reasonable combinations:) - with all distances from glycosite

comparisons = [('likely pathogenic + pathogenic','benign'),
                ('pathogenic','benign'),
                ('pathogenic','likely benign + benign'),
                ('likely pathogenic + pathogenic','likely benign + benign')]

for comp in comparisons:
    i,j = comp
    
    contingency_table = counts_pivot.loc[[i,j]]
    oddratio,p = scipy.stats.fisher_exact(contingency_table)
    if p < 0.1:
        print(contingency_table)
        print(oddratio,p)
        print('\n')

## Nothing close to significant. 

Note, not shown here but nothing close to significant if I filter for JUST residues within 6Angstrom of a glycosite.

## What if I segregate by amino acid substitution type?

by_aas = df.groupby(by=['glycosite_proximal','clinvar_pathogenic','aa_change',],as_index=False).count()[['glycosite_proximal','clinvar_pathogenic','aa_change','#chr']]

by_aas = by_aas.sort_values(by='aa_change')
by_aas

counts_pivot = counts.pivot(index='clinvar_pathogenic',columns='glycosite_proximal',values='#chr')
counts_pivot.loc['likely benign + benign'] = counts_pivot.loc['likely benign'] + counts_pivot.loc['benign'] 
counts_pivot.loc['likely pathogenic + pathogenic'] = counts_pivot.loc['likely pathogenic'] + counts_pivot.loc['pathogenic'] 

for aa in by_aas['aa_change'].unique():
    
    one_aa_change = by_aas[by_aas['aa_change'] == aa]
    counts_pivot = one_aa_change.pivot(index='clinvar_pathogenic',columns='glycosite_proximal',values='#chr')
    
    # Sometime AAs will not have all the mutation categories, making sure they have them here:
    expected_rows = ['benign','likely benign','likely pathogenic','pathogenic']
    for r in expected_rows:
        if r not in counts_pivot.index:
            counts_pivot.loc[r] = 0
    
    counts_pivot.loc['likely benign + benign'] = counts_pivot.loc['likely benign'] + counts_pivot.loc['benign'] 
    counts_pivot.loc['likely pathogenic + pathogenic'] = counts_pivot.loc['likely pathogenic'] + counts_pivot.loc['pathogenic'] 
    counts_pivot = counts_pivot.fillna(0)
    
    for comp in comparisons:
        i,j = comp

        contingency_table = counts_pivot.loc[[i,j]]

        if (contingency_table.shape[0] != 2) or (contingency_table.shape[1] != 2):
            continue

        oddratio,p = scipy.stats.fisher_exact(contingency_table)
        if p < 0.05:
            print('{}\tp={}'.format(aa,p))
            print(contingency_table)
            print('\n')
