==> code/04_clean_protein_continuous_dist_table.ipynb <==
import pandas as pd
import numpy as np

# PDB
#df = pd.read_csv('../data/GTspec_pdb_query.protein_structure_info.just_glyco.csv').drop_duplicates()
# Swiss
df = pd.read_csv('../data/GTspec_swissmod_query.protein_structure_info.just_glyco.csv').drop_duplicates()
df = df.dropna(subset=['struct_window_angstrom_resnum'])

df[['gene_id','angstroms_to_feature','struct_residue']]

print(df.shape)
df = df[df['struct_residue'].isin(['N'])]
print(df.shape)

## N glycosylation sites may be easier to evaluate 
- S, T are O-glycosylation
- Anything that is not S, T or N should be filtered

proteins = []
aas = []
distances = []
for i in df.index:
    distance_and_aa_list = df.loc[i]['angstroms_to_feature'].split('|')
    aa_list = [da.split('_')[-1] for da in distance_and_aa_list]
    distance_list = [da.split('_')[0] for da in distance_and_aa_list]
    protein = [df.loc[i]['gene_id']] * len(distance_list)
    
    aas.extend(aa_list)
    distances.extend(distance_list)
    proteins.extend(protein)

merged = pd.concat([pd.Series(proteins),
           pd.Series(aas,dtype=int),
           pd.Series(distances,dtype=float)],
           axis=1)

merged.columns = ['protein','aa','distance_to_feature']

## If multiple features per amino acid, keep the closest distance one

print(merged.shape)
merged = merged.sort_values(by='distance_to_feature').drop_duplicates(subset=['protein','aa'],keep='first')
print(merged.shape)

merged.to_csv('../data/protein_aa_distance.txt',sep='\t',header=None)
