==> code/08_pathogenicity_software_merged_with_glycoimpact.ipynb <==
import pandas as pd

## Reading in dbNSFP data

db = pd.read_csv('../data/dbNSFP3.2a_gprot_simple_aa_uniq_w_header.txt',
                 sep='\t',
                 index_col=False)

print(db.shape)

software = db[db.columns[db.columns.str.contains('rank')]].columns.to_list()

db['aa_change'] = db['aaref'] + '=>' + db['aaalt']

db = db[['aapos','aa_change','Uniprot_acc_Polyphen2','clinvar_clnsig'] + software]
db = db.drop_duplicates(subset=['aapos','aa_change','Uniprot_acc_Polyphen2'])
db.head()

print(db.shape)

## Create simple table for just proteins with glycosylation sites with distance annotation

p_d = pd.read_csv('../data/protein_aa_distance.txt',
                       sep='\t',
                       header=None,index_col=0)

p_d.columns = ['proteins','aa','distance_to_feature']

## Going through the clinvar database and matching amino acids


pd_dict = {'aa':[],
           'proteins':[],
           'clinvar_path':[],
           'aa_change':[],
           'clinvar_index':[]}
           
for s in software:
    pd_dict[s] = []

for i in db.index:
    one_var = db.loc[i]
    aa_list = one_var['aapos']
    if not isinstance(aa_list,int):
        aa_list = one_var['aapos'].split(';')
    else:
        aa_list = [one_var['aapos']]
    protein_list = one_var['Uniprot_acc_Polyphen2'].split(';')
    path = one_var['clinvar_clnsig']
    aa_change = one_var['aa_change']
    for p in protein_list:
        for aa in aa_list:
            pd_dict['aa'].append(aa)
            pd_dict['proteins'].append(p)
            pd_dict['clinvar_path'].append(path)
            pd_dict['aa_change'].append(aa_change)
            pd_dict['clinvar_index'].append(i)
            for s in software:
                pd_dict[s].append(one_var[s])

flat_db = pd.DataFrame.from_dict(pd_dict)
flat_db['aa'] = flat_db['aa'].astype(int)

print(flat_db.shape)
print('Dropping duplicates')
flat_db =  flat_db.drop_duplicates(subset=['aa','proteins','aa_change'])
print(flat_db.shape)

p_d.head()

flat_db.head()

merged = flat_db.merge(p_d,on=['aa','proteins'],how='outer')

print(merged.shape)

# Dropping any variants that are not in glycosylated proteins

g_proteins = list(p_d['proteins'].unique())

merged = merged[merged['proteins'].isin(g_proteins)]
print(merged.shape)

# Dealing with variants that do not have a distance to glycosylation site
Need to be minful here: can either filter these, add psudo distance or keep as background


## Dropping NAs
print(merged.shape)
merged = merged[~merged['distance_to_feature'].isna()]
print(merged.shape)

## Filtering sites where distance to feature == 0, which is presumably the glycosylation site

merged = merged[merged['distance_to_feature'] != 0]
merged.shape

## Dropping duplicates that have same aa, aapos and , keeping closest to glycosite

merged = merged.sort_values(by='distance_to_feature',ascending=True).drop_duplicates(subset=['aa','proteins','aa_change'],keep='first')

print(merged.shape)

## Dropping variants that don't have at least 4 scores

merged = merged.dropna(subset=software,thresh=3)
print(merged.shape)

merged.shape

## Adding in glycoimpact score

gi = pd.read_csv('../data/glyco_protein_impact_within_6.csv')

gi['aa_change'] = gi['substitution']

merged = merged.merge(gi,on='aa_change',how='outer').dropna(subset=['aa'])

merged.to_csv('../data/glycoproteins_distances_with_pathsoftware.txt',sep='\t')
